+++ maas-20260318/alibabacloud_maas20260318/models/_create_api_key_request.py
@@ -11,7 +11,6 @@ class CreateApiKeyRequest(DaraModel):
         workspace_id: str = None,
     ):
         self.description = description
-        # This parameter is required.
         self.workspace_id = workspace_id
 
     def validate(self):