+++ maas-20260318/alibabacloud_maas20260318/models/_update_api_key_request.py
@@ -7,13 +7,8 @@ from darabonba.model import DaraModel
 class UpdateApiKeyRequest(DaraModel):
     def __init__(
         self,
-        api_key_id: int = None,
         description: str = None,
     ):
-        # API Key ID。
-        # 
-        # This parameter is required.
-        self.api_key_id = api_key_id
         self.description = description
 
     def validate(self):
@@ -24,9 +19,6 @@ class UpdateApiKeyRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
-        if self.api_key_id is not None:
-            result['apiKeyId'] = self.api_key_id
-
         if self.description is not None:
             result['description'] = self.description
 
@@ -34,9 +26,6 @@ class UpdateApiKeyRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
-        if m.get('apiKeyId') is not None:
-            self.api_key_id = m.get('apiKeyId')
-
         if m.get('description') is not None:
             self.description = m.get('description')