+++ maas-20260318/ChangeLog.md
@@ -1,3 +1,9 @@
+2026-03-26 Version: 2.0.0
+- Update API DeleteApiKey: change request The number of path parameters has changed.
+- Update API GetApiKey: change request The number of path parameters has changed.
+- Update API UpdateApiKey: change request The number of path parameters has changed.
+
+
 2026-03-26 Version: 1.1.0
 - Support API ListWorkspaces.